backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dataclasses-json==0.6.1
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.65.2
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.343
langchain-core==0.0.7
langsmith==0.0.67
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
pydantic==1.10.13
pymongo==4.5.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.14.2
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.14.0
yarl==1.9.2
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-compute==1.19.2
googleapis-common-protos==1.63.2